YAML loading for SimpleRulebook (take 2). (#8575)
* YAML loading for SimpleRulebook.

Also, provide SimpleRulebook with a RemoveAll() method.

This commit was originally reviewed/merged as PR #8525,
but that broke the mac build and was reverted in #8566.

* Changes to fix the mac build.

Specifically:  make the exception expectations in unit-tests
less specific/brittle.  (The mac build failed due to a different
yaml-cpp version that throws exceptions with message text.)